+++ kurtosis-lambda/impl/ethereum_kurtosis_lambda_args.go
@@ -1,6 +1,0 @@
-package impl
-
-type EthereumKurtosisLambdaArgs struct {
-	// Indicates the log level for this Kurtosis Lambda implementation
-	LogLevel string `json:"logLevel"`
-}