+++ docs/changelog.md
@@ -1,4 +1,8 @@
 # TBD
+### Changes
+* Start only 2 child ETH nodes (for 3 total, with the boot node) rather than 4
+* Start all child ETH nodes then wait for them all to become available, rather than doing "start first node, wait for it, start second node, wait for it..."
+* Changed the result object to also contain local host machine port bindings, so users can easily access the cluster in interactvie/debug mode
 
 # 0.1.4
 ### Fixes